+++ srcs/commands/command_var_extension.c
@@ -6,7 +6,7 @@
 /*   By: paulorod <paulorod@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/04 15:14:29 by paulorod          #+#    #+#             */
-/*   Updated: 2023/09/11 16:02:28 by paulorod         ###   ########.fr       */
+/*   Updated: 2023/09/12 15:50:25 by paulorod         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
